requirements.txt
fastapi
llama_cpp_python
transformers
pydantic
fastapi
asyncer
requests